refactor: factored out ParameterReadingSlider
